Rework RMQQueue#subscribe and basic.consume API
Due to how options structures and specifically RMQBasicConsumeOptions
are serialised, the ManualAcknowledgements variant only worked by accident
(noLocal is not supported but default acknowledgement mode is manual).

There is no way to rename the unfortunate noAck option without
modifying the protocol source document or special casing that
method in the code generator.

So instead we revert the RMQBasicConsumeOptions change and modify
RMQChannel and RMQQueue API to look better and allow for more flexibility.

For example,

let cons = ch.queue("", options: [.exclusive])
            .bind(x)
            .subscribe(withAckMode: [.auto]) { _ in }

looks very similar to the original RMQBasicConsumeOptions
interface in Swift, and additional selectors make it even more
clear in Objective-C.

When more than one option is required, RMQBasicConsumeOptions
remains the only option. This may be further improved
in the future.