Avoid RMQConnection leaks by adopting weak refs, take 2 (on behalf of @BarryDuggan)
for some important dependencies of allocated channels and
RMQConnection itself.

In this client, RMQConnection auto-allocates a channel for
the purpose of special "channel zero" (system communication
in the protocol) purposes, and that leads to a loop
of strong references that prevent RMQConnection instances
from being released.

Contributed by @BarryDuggan in #194.